[advanced] 유틸함수 formatWeek, formatMonth 분리
