+++ packages/assignment-7/src/hooks/useEvents.ts
@@ -359,19 +359,6 @@ function useEvents(): UseEventsReturn {
     });
   })();
 
-  const formatWeek = (date: Date): string => {
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const weekNumber = Math.ceil(date.getDate() / 7);
-    return `${year}년 ${month}월 ${weekNumber}주`;
-  };
-
-  const formatMonth = (date: Date): string => {
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    return `${year}년 ${month}월`;
-  };
-
   return {
     fetchHolidays,
     fetchEvents,
@@ -388,8 +375,6 @@ function useEvents(): UseEventsReturn {
     navigate,
     searchEvents,
     filteredEvents,
-    formatWeek,
-    formatMonth,
     events,
     setEvents,
     title,
