+++ packages/assignment-7/src/utils.ts
@@ -1,0 +1,12 @@
+export const formatWeek = (date: Date): string => {
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const weekNumber = Math.ceil(date.getDate() / 7);
+  return `${year}년 ${month}월 ${weekNumber}주`;
+};
+
+export const formatMonth = (date: Date): string => {
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  return `${year}년 ${month}월`;
+};
